File encryption support

+++ Android/src/bencoding/securely/FileCryptoProxy.java
@@ -1,8 +1,13 @@
 package bencoding.securely;
 
-import java.io.File;
+
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.util.HashMap;
 
@@ -10,6 +15,7 @@ import javax.crypto.Cipher;
 import javax.crypto.CipherInputStream;
 import javax.crypto.CipherOutputStream;
 import javax.crypto.KeyGenerator;
+import javax.crypto.NoSuchPaddingException;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 
@@ -17,166 +23,215 @@ import org.appcelerator.kroll.KrollDict;
 import org.appcelerator.kroll.KrollFunction;
 import org.appcelerator.kroll.KrollProxy;
 import org.appcelerator.kroll.annotations.Kroll;
+import org.appcelerator.titanium.TiApplication;
 import org.appcelerator.titanium.TiC;
+import org.appcelerator.titanium.TiLifecycle;
 import org.appcelerator.titanium.io.TiBaseFile;
 import org.appcelerator.titanium.io.TiFileFactory;
 import org.appcelerator.titanium.util.TiConvert;
 
-@Kroll.proxy(creatableInModule=SecurelyModule.class)
-public class FileCryptoProxy extends KrollProxy {
+import android.app.Activity;
 
+@Kroll.proxy(creatableInModule=SecurelyModule.class)
+public class FileCryptoProxy  extends KrollProxy implements TiLifecycle.OnLifecycleEvent 
+{
+
+	public static final String RESOURCE_ROOT_ASSETS = "file:///android_asset/Resources";
 	static int _aesBytes = 128;
-	
+
 	public FileCryptoProxy()
 	{
-		super();		
+		super();
+
 	}
 	
-	private class AESFilesRunnable implements Runnable
+	private boolean isInResources(TiBaseFile file)
+	{		
+		if (file.isFile()) {
+			return file.nativePath().startsWith(RESOURCE_ROOT_ASSETS);
+		}
+		return false;
+	}
+	
+	private void copy(InputStream is, OutputStream os) throws IOException {
+	    int i;
+	    byte[] b = new byte[1024];
+	    while((i=is.read(b))!=-1) {
+	      os.write(b, 0, i);
+	    }
+	  }
+
+	private static byte[] getRawKey(byte[] seed) throws Exception {
+		KeyGenerator kgen = KeyGenerator.getInstance("AES");
+		SecureRandom sr = SecureRandom.getInstance("SHA1PRNG", "Crypto");
+		sr.setSeed(seed);
+		try {
+			kgen.init(256, sr);
+			} catch (Exception e) {
+			// Log.w(LOG, "This device doesn't support 256 bits, trying 192 bits.");
+			try {
+			kgen.init(192, sr);
+			} catch (Exception e1) {
+			// Log.w(LOG, "This device doesn't support 192 bits, trying 128 bits.");
+			kgen.init(128, sr);
+			}
+		}
+		SecretKey skey = kgen.generateKey();
+		byte[] raw = skey.getEncoded();
+		return raw;
+	}
+    private void doCallback(KrollFunction callback,HashMap<String, Object> event){
+		if(callback!=null){		
+			callback.call(getKrollObject(), event);
+		}		    	
+    }
+
+	private class AESDecryptRunnable implements Runnable
 	{
-		private boolean _isEncrypt = true;
-		private File _to = null;
-		private File _from = null;
+		private FileOutputStream _to = null;
+		private FileInputStream _from = null;
 		KrollFunction _callback = null;
 		String _secret = null;
-		
-		public AESFilesRunnable(String secret, KrollFunction callback, boolean isEncrypt, File from, File to){
-			_isEncrypt = isEncrypt;
-			_to = to;
+		String _inputFile = null;
+		String _outputFile = null;
+		public AESDecryptRunnable(String password, 
+								 FileInputStream from, 
+								 FileOutputStream to,
+								 KrollFunction callback,
+								 String inputFile, String outputFile ){
+			_secret = password;
+			_to=to;
 			_from = from;
-			_callback = callback;
-			_secret = secret;
-		}
-		private byte[] getRawKey(byte[] seed) throws Exception {
-		    KeyGenerator kgen = KeyGenerator.getInstance("AES");
-		    SecureRandom sr = SecureRandom.getInstance("SHA1PRNG");
-		    sr.setSeed(seed);
-		    kgen.init(_aesBytes, sr); // 192 and 256 bits may not be available
-		    SecretKey skey = kgen.generateKey();
-		    byte[] raw = skey.getEncoded();
-		    return raw;
+			_callback=callback;
+			_inputFile = inputFile;
+			_outputFile = outputFile;			
 		}
 		
-		private Cipher getCipherEncrypt(byte[] raw) throws Exception {
-		    SecretKeySpec skeySpec = new SecretKeySpec(raw, "AES");
-		    Cipher cipher = Cipher.getInstance("AES");
-		    cipher.init(Cipher.ENCRYPT_MODE, skeySpec);
-		    return cipher;
+		HashMap<String, Object> buildResult(Boolean success){
+			HashMap<String, Object> event = new HashMap<String, Object>();
+			event.put(TiC.PROPERTY_SUCCESS, success);	
+			event.put("to",_outputFile);	
+			event.put("from",_inputFile);
+			return event;
 		}
-		private Cipher getCipherDecrypt(byte[] raw) throws Exception {
-		    SecretKeySpec skeySpec = new SecretKeySpec(raw, "AES");
-		    Cipher cipher = Cipher.getInstance("AES");
-		    cipher.init(Cipher.DECRYPT_MODE, skeySpec);
-		    return cipher;
-		}		
-//		 private byte[] getKeyBytes(final byte[] key) throws Exception {
-//		        byte[] keyBytes = new byte[16];
-//		        System.arraycopy(key, 0, keyBytes, 0, Math.min(key.length, keyBytes.length));
-//		        return keyBytes;
-//		    }
-//		    public Cipher getCipherEncrypt(final byte[] key) throws Exception {
-//		        byte[] keyBytes = getKeyBytes(key);
-//		        Cipher cipher = Cipher.getInstance("AES/CBC/PKCS5Padding");
-//		        SecretKeySpec secretKeySpec = new SecretKeySpec(keyBytes, "AES");
-//		        IvParameterSpec ivParameterSpec = new IvParameterSpec(keyBytes);
-//		        cipher.init(Cipher.ENCRYPT_MODE, secretKeySpec, ivParameterSpec);
-//		        return cipher;
-//		    }
-//		    public Cipher getCipherDecrypt(byte[] key) throws Exception {
-//		        byte[] keyBytes = getKeyBytes(key);
-//		        Cipher cipher = Cipher.getInstance("AES/CBC/PKCS5Padding");
-//		        SecretKeySpec secretKeySpec = new SecretKeySpec(keyBytes, "AES");
-//		        IvParameterSpec ivParameterSpec = new IvParameterSpec(keyBytes);
-//		        cipher.init(Cipher.DECRYPT_MODE, secretKeySpec, ivParameterSpec);
-//		        return cipher;
-//		    }		
-		    public void encrypt(File inputFile, File outputFile, byte[] key) throws Exception {
-		        Cipher cipher = getCipherEncrypt(key);
-		        FileOutputStream fos = null;
-		        CipherOutputStream cos = null;
-		        FileInputStream fis = null;
-		        try {
-		        	fis = new FileInputStream(inputFile);
-		            fos = new FileOutputStream(outputFile);
-		            cos = new CipherOutputStream(fos, cipher);
-		            fos = null;
-		            byte[] data = new byte[1024];
-		            int read = fis.read(data);
-		            while (read != -1) {
-		                cos.write(data, 0, read);
-		                read = fis.read(data);
-		                System.out.println(new String(data, "UTF-8").trim());
-		            }
-		            cos.flush();
-		        } finally {
-		            if (cos != null) {
-		                cos.close();
-		             }
-		             if (fos != null) {
-		                fos.close();
-		             }
-		             if (fis != null) {
-		                fis.close();
-		             }
-		        }
-		    }	
-		    public void decrypt(File inputFile, File outputFile, byte[] key) throws Exception {
-		        Cipher cipher = getCipherDecrypt(key);
-		        FileOutputStream fos = null;
-		        CipherInputStream cis = null;
-		        FileInputStream fis = null;
-		        try {
-		            fis = new FileInputStream(inputFile);
-		            cis = new CipherInputStream(fis, cipher);
-		            fos = new FileOutputStream(outputFile);
-		            byte[] data = new byte[1024];
-		            int read = cis.read(data);
-		            while (read != -1) {
-		                fos.write(data, 0, read);
-		                read = cis.read(data);
-		                System.out.println(new String(data, "UTF-8").trim());
-		            }
-		        } finally {
-		            fos.close();
-		            cis.close();
-		            fis.close();
-		        }
-		    }		
-		    
-		    void doCallback(boolean success,String msg){
-				if(_callback!=null){
-	    			HashMap<String, Object> event = new HashMap<String, Object>();
-	    			event.put(TiC.PROPERTY_SUCCESS, success);
-	    			event.put("message", msg);
-	    			event.put("to",_to.getAbsolutePath());	
-	    			event.put("from",_from.getAbsolutePath());			
-					_callback.call(getKrollObject(), event);
-				}		    	
-		    }
-			@Override
-			public void run() {
-				try {	
-					
-					if(_isEncrypt){
-						encrypt(_from, _to, getRawKey(_secret.getBytes()));			
-					}else{
-						decrypt(_from, _to, getRawKey(_secret.getBytes()));
-					}
-					doCallback(true,"completed");
+		@Override
+		public void run() {
+			try {	
+				byte[] raw = null;
+				raw = getRawKey(_secret.getBytes());				
+				SecretKeySpec skeySpec = new SecretKeySpec(raw, "AES");
+				Cipher cipher = Cipher.getInstance("AES");
+				cipher.init(Cipher.DECRYPT_MODE, skeySpec);					
+				CipherInputStream is = new CipherInputStream(_from,cipher);
 				
-				} catch (Exception e) {
-					e.printStackTrace();
-					doCallback(false,e.toString());
-				}finally{
-					if(_to!=null){
-						_to=null;
-					}
-					if(_from!=null){
-						_from=null;
-					}					
+				copy(is, _to);
+
+				is.close();
+				_to.close();
+				_from.close();						
+				doCallback(_callback, buildResult(true));
+								
+			} catch (NoSuchAlgorithmException e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);					
+			} catch (NoSuchPaddingException e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);
+			} catch (Exception e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);			
+			}finally{
+				if(_to!=null){
+					_to=null;
 				}
-			}		
+				if(_from!=null){
+					_from=null;
+				}					
+			}
+		}			
 	}
+	private class AESEncryptRunnable implements Runnable
+	{
+		private FileOutputStream _to = null;
+		private FileInputStream _from = null;
+		KrollFunction _callback = null;
+		String _secret = null;
+		String _inputFile = null;
+		String _outputFile = null;
+		public AESEncryptRunnable(String password, 
+								 FileInputStream from, 
+								 FileOutputStream to,
+								 KrollFunction callback,
+								 String inputFile, String outputFile ){
+			_secret = password;
+			_to=to;
+			_from = from;
+			_callback=callback;
+			_inputFile = inputFile;
+			_outputFile = outputFile;			
+		}
+		
+		HashMap<String, Object> buildResult(Boolean success){
+			HashMap<String, Object> event = new HashMap<String, Object>();
+			event.put(TiC.PROPERTY_SUCCESS, success);	
+			event.put("to",_outputFile);	
+			event.put("from",_inputFile);
+			return event;
+		}
+		@Override
+		public void run() {
+			try {	
+
+				byte[] raw = null;
+				raw = getRawKey(_secret.getBytes());
+
+				SecretKeySpec skeySpec = new SecretKeySpec(raw, "AES");
+				Cipher cipher = Cipher.getInstance("AES");
+				cipher.init(Cipher.ENCRYPT_MODE, skeySpec);				
+				CipherOutputStream os = new CipherOutputStream(_to, cipher);
+
+				copy(_from, os);
+				os.close();
+				doCallback(_callback, buildResult(true));	
+				
+			} catch (NoSuchAlgorithmException e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);					
+			} catch (NoSuchPaddingException e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);
+			} catch (Exception e) {
+				e.printStackTrace();
+				LogHelpers.Log(e);
+				HashMap<String, Object> event = buildResult(false);
+				event.put("message",e.toString());	
+				doCallback(_callback, event);	
+			}finally{
+				if(_to!=null){
+					_to=null;
+				}
+				if(_from!=null){
+					_from=null;
+				}					
+			}
+		}			
+	}
+	
 
 	@Override
 	public void handleCreationDict(KrollDict options)
@@ -194,8 +249,8 @@ public class FileCryptoProxy extends KrollProxy {
 		if(!args.containsKey("completed")){
 			throw new IllegalArgumentException("missing completed callback method");
 		}
-		if(!args.containsKey("secret")){
-			throw new IllegalArgumentException("missing secret");
+		if(!args.containsKey("password")){
+			throw new IllegalArgumentException("missing password");
 		}
 		if(!args.containsKey("to")){
 			throw new IllegalArgumentException("missing output file");
@@ -203,7 +258,7 @@ public class FileCryptoProxy extends KrollProxy {
 		if(!args.containsKey("from")){
 			throw new IllegalArgumentException("missing file to decrypt");
 		}		
-		String secret = args.getString("secret");
+		String secret = args.getString("password");
 		KrollFunction callback = null;
 		Object object = args.get("completed");
 		if (object instanceof KrollFunction) {
@@ -216,25 +271,64 @@ public class FileCryptoProxy extends KrollProxy {
 		}
 		String rawOutputFile = args.getString("to");
 		TiBaseFile outputFile = TiFileFactory.createTitaniumFile(rawOutputFile,false);
+		if(isInResources(outputFile)){
+			throw new IllegalArgumentException("Output file cannot be in a protected directory " + outputFile.nativePath());
+		}
 		if(outputFile.exists()){
 			outputFile.deleteFile();
 		}
+  
+		try{
+			
+			FileInputStream fis = null;
+			if(isInResources(sourceFile)){
+				String path = removeResourcesDirectory(sourceFile.nativePath());
+				fis = new FileInputStream(TiApplication.getInstance().getApplicationContext().getFileStreamPath(path));
+			}else{
+				fis = new FileInputStream(sourceFile.getNativeFile());
+			}
 
-		Thread clientThread = new Thread(new AESFilesRunnable(secret,callback,false,sourceFile.getNativeFile(),outputFile.getNativeFile()));
-		clientThread.run();
-        
+			String path = removeFilePrefix(outputFile.nativePath());
+			//LogHelpers.Log("path=" + path);					
+	        FileOutputStream fos = new FileOutputStream(path); 
+	        			
+			Thread clientThread = new Thread(new AESDecryptRunnable(secret,fis,fos, callback,sourceFile.nativePath(),outputFile.nativePath() ));
+			clientThread.run();
+			
+		} catch (FileNotFoundException e) {
+			e.printStackTrace();
+			LogHelpers.Log(e);
+		}
 	}
 
+	private String removeResourcesDirectory(String path){
+		if(path.startsWith("file:///android_asset/Resources/")){
+			int len = "file:///android_asset/Resources/".length();
+			return path.substring(len);
+		}else{
+			return path;
+		}
+	}
+   
+	private String removeFilePrefix(String path){
+		if(path.startsWith("file:/")){
+			int len = "file:/".length();
+			return path.substring(len);
+		}else{
+			return path;
+		}
+	}
+	
 	@SuppressWarnings({ "rawtypes", "unchecked" })
 	@Kroll.method
 	public void AESEncrypt(HashMap hm){
 		KrollDict args = new KrollDict(hm);
-		
+
 		if(!args.containsKey("completed")){
 			throw new IllegalArgumentException("missing completed callback method");
 		}
-		if(!args.containsKey("secret")){
-			throw new IllegalArgumentException("missing secret");
+		if(!args.containsKey("password")){
+			throw new IllegalArgumentException("missing password");
 		}
 		if(!args.containsKey("to")){
 			throw new IllegalArgumentException("missing output file");
@@ -242,7 +336,7 @@ public class FileCryptoProxy extends KrollProxy {
 		if(!args.containsKey("from")){
 			throw new IllegalArgumentException("missing file to decrypt");
 		}		
-		String secret = args.getString("secret");
+		String secret = args.getString("password");
 		KrollFunction callback = null;
 		Object object = args.get("completed");
 		if (object instanceof KrollFunction) {
@@ -253,13 +347,47 @@ public class FileCryptoProxy extends KrollProxy {
 		if(!sourceFile.exists()){
 			throw new IllegalArgumentException("File to encrypt does not exist " + sourceFile.nativePath());
 		}
-		
+
 		String rawOutputFile = args.getString("to");
 		TiBaseFile outputFile = TiFileFactory.createTitaniumFile(rawOutputFile,false);
+		if(isInResources(outputFile)){
+			throw new IllegalArgumentException("Output file cannot be in a protected directory " + outputFile.nativePath());
+		}
+		
 		if(outputFile.exists()){
 			outputFile.deleteFile();
 		}
-		Thread clientThread = new Thread(new AESFilesRunnable(secret,callback,true,sourceFile.getNativeFile(),outputFile.getNativeFile()));
-		clientThread.run();
+
+		try{
+				
+			FileInputStream fis = null;
+			if(isInResources(sourceFile)){
+				String path = removeResourcesDirectory(sourceFile.nativePath());
+				fis = new FileInputStream(TiApplication.getInstance().getApplicationContext().getFileStreamPath(path));
+			}else{
+				fis = new FileInputStream(sourceFile.getNativeFile());
+			}
+
+			String path = removeFilePrefix(outputFile.nativePath());
+			//LogHelpers.Log("path=" + path);					
+	        FileOutputStream fos = new FileOutputStream(path);
+			
+			Thread clientThread = new Thread(new AESEncryptRunnable(secret,fis,fos, callback,sourceFile.nativePath(),outputFile.nativePath() ));
+			clientThread.run();
+			
+		} catch (FileNotFoundException e) {
+			e.printStackTrace();
+			LogHelpers.Log(e);			
+		}	
 	}
+	@Override
+	public void onDestroy(Activity arg0) {}
+	@Override
+	public void onPause(Activity arg0) {}
+	@Override
+	public void onResume(Activity arg0) {}
+	@Override
+	public void onStart(Activity arg0) {}
+	@Override
+	public void onStop(Activity arg0) {}
 }
